在原有的字段上拓展了一个functions字段，优先级低于tools。
